+++ src/desktop/Header/Header.module.css
@@ -41,6 +41,7 @@
 .logo {
     display: flex;
     align-items: center;
+    text-decoration: none;
 }
 
 .logoText {
@@ -115,36 +116,44 @@
     border-radius: 10px;
     background: white;
     box-shadow: 0px 0px 4px 5px rgba(0,0,0,.04);
-    transform-origin:top;
-    transform: translateY(-2px);
-    pointer-events: none;
-    opacity: 0;
+    display: flex;
+    flex-direction: column;
 }
 
-.searchResultsShown {
-    pointer-events: auto;
-    opacity: 1;
-    transform: translateY(0);
+.searchResults a {
+    text-decoration: none;
+}
+
+.searchResults a:last-child {
+    border-top:2px solid var(--border-color);
+    color:black;
+    font-weight: 600;
+    text-align: center;
 }
 
 /* search results item */
 .searchResultsItem {
-    height: 50px;
-    display: flex;
-    align-items: center;
+    flex-grow: 3;
+    line-height: 50px;
     padding: 0 15px;
+    transition: .3s;
+    color:var(--grey-color);
 }
 
+.searchResultsItem:hover {
+    background:var(--border-color);
+}
 
 /* info button */
 .info {
     padding: 0 20px;
 }
 
-.info span {
-    cursor: pointer;
+.info a {
     border-radius:50%;
     border: 1px solid var(--border-color);
+    text-decoration: none;
+    color:var(--grey-color);
     display: flex;
     align-items: center;
     justify-content: center;
@@ -186,10 +195,26 @@
     border-radius: 10px;
     text-align: center;
     overflow: hidden;
+
+    opacity: 0;
+    transform: translateY(-10px);
+    pointer-events: none;
+    transition: .3s;
+}
+
+/* active */
+.languageAllActive {
+    opacity: 1;
+    transform: translateY(0);
+    pointer-events: auto
+}
+/* end active */
+
+.languageAll a {
+    text-decoration: none;
 }
 
 .languageAll div {
-    cursor: pointer;
     text-transform: uppercase;
     color:var(--grey-color);
     padding: 5px;
