--- src/stories/Header/Header.stories.ts
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Header } from './Header';
-
-const meta: Meta<typeof Header> = {
-    title: 'header',
-  component: Header,
-};
-
-export default meta;
-type Story = StoryObj<typeof Header>;
-
-export const Basic: Story = {};
