--- src/stories/TestComponent.stories.ts
@@ -1,21 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { TestComponent } from '../index';
-
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
-const meta = {
-  title: 'Example/TestComponent',
-  component: TestComponent,
-  tags: ['autodocs'],
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
-} satisfies Meta<typeof TestComponent>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Primary: Story = {
-  args: {},
-};
